implement preprocessing.process_scan, add logging
